+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/GetProvider/GetProviderQueryRequest.cs
@@ -4,6 +4,6 @@ namespace SFA.DAS.EAS.Application.Queries.GetProvider
 {
     public class GetProviderQueryRequest : IAsyncRequest<GetProviderQueryResponse>
     {
-        public int ProviderId { get; set; }
+        public long ProviderId { get; set; }
     }
 }
